+++ CHANGELOG.md
@@ -2,7 +2,10 @@ Dagon 0.33.0 - TBD
 ------------------
 - **Core**
   - Reworked settings.conf handling code in `Application`. Now each VFS entry can have its own settings.conf file overriding a default one in the executable directory. Also default settings.conf is now loaded before initializing other systems, which allows, in particular, to override logger settings from the beginning
-  - New config options: `logLevel`, `logToStdout`, `logFile`, `logTimestampTags`, `logLevelTags`, `windowResizable`, `windowX`, `windowY`, `vsync`, `glDebugOutput`, `SDL2Path`, `SDL2ImagePath`.
+  - Breaking change: renderer and post-processing settings are now in render.conf, `enableShaderCache` is now `gl.enableShaderCache`
+  - New settings.conf options: `log.level`, `log.toStdout`, `log.file`, `log.timestampTags`, `log.levelTags`, `window.resizable`, `window.x`, `window.y`, `vsync`, `gl.debugOutput`, `SDL2.path`, `SDL2Image.path`
+- **Extensions**
+  - Breaking change: `dagon:audio` settings are now in audio.conf.
 
 Dagon 0.32.1 - 20 Oct, 2025
 ---------------------------